requirements.txt
streamlit
sqlalchemy
pandas
google-genai
groq
python-dotenv
pycryptodome
plotly
